multiview: Move Views panel to RenderLayer panel
