+++ blender/source/blender/editors/space_buttons/buttons_header.c
@@ -141,7 +141,6 @@ void buttons_header_buttons(const bContext *C, ARegion *ar)
 
 	BUTTON_HEADER_CTX(BCONTEXT_RENDER, ICON_SCENE, N_("Render"));
 	BUTTON_HEADER_CTX(BCONTEXT_RENDER_LAYER, ICON_RENDERLAYERS, N_("Render Layers"));
-	BUTTON_HEADER_CTX(BCONTEXT_RENDER_VIEW, ICON_RENDERVIEWS, N_("Render Views"));
 	BUTTON_HEADER_CTX(BCONTEXT_SCENE, ICON_SCENE_DATA, N_("Scene"));
 	BUTTON_HEADER_CTX(BCONTEXT_WORLD, ICON_WORLD, N_("World"));
 	BUTTON_HEADER_CTX(BCONTEXT_OBJECT, ICON_OBJECT_DATA, N_("Object"));
