+++ blender/source/blender/makesdna/DNA_space_types.h
@@ -184,7 +184,6 @@ typedef enum eSpaceButtons_Context {
 	BCONTEXT_CONSTRAINT = 11,
 	BCONTEXT_BONE_CONSTRAINT = 12,
 	BCONTEXT_RENDER_LAYER = 13,
-	BCONTEXT_RENDER_VIEW = 14,
 	
 	/* always as last... */
 	BCONTEXT_TOT
